+++ thaw/src/input/mod.rs
@@ -62,8 +62,8 @@ pub fn Input(
     /// Size of the input (changes the font size and spacing).
     #[prop(optional, into)]
     size: Signal<InputSize>,
+    #[prop(optional, into)] autocomplete: MaybeProp<String>,
 ) -> impl IntoView {
-    let input_style = move || input_style.get().unwrap_or_default();
     mount_style("input", include_str!("./input.css"));
     let (id, name) = FieldInjection::use_id_and_name(id, name);
     let validate = Rule::validate(rules, value, name);
@@ -185,7 +185,8 @@ pub fn Input(
                 size=input_size
                 placeholder=move || placeholder.get()
                 node_ref=input_ref
-                style=input_style
+                style=move || input_style.get()
+                autocomplete=move || autocomplete.get()
             />
 
             {if let Some(suffix) = input_suffix.and_then(|suffix| suffix.if_.then_some(suffix)) {